# doc/api.yaml
openapi: 3.0.0
info:
  title: WASA Photo
  x-summary: The api that allows to use WASA Photo backend
  description: |-
    This API allows using WASA Photo with a variety of
    functionalities like image upload, user profile photos,
    stream and information etc.
  version: 1.0.0
...
servers:
  - url: '{protocol}://{host}:{port}'
    description: Applcation server, use this parameters for local development and production
    x-sandbox: true
    variables:
      protocol:
        enum:
          - http
          - https
        default: http
      host:
        default: localhost
        description: Application server's host
      port:
        enum:
          - '3000'
          - '443'
          - '80'
        default: '3000'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: accessToken
      description: >-
        Enter the token with the `Bearer: ` prefix, e.g. "Bearer abcde12345".
  responses:
...
    UnauthorizedError:
      description: Access token is missing or invalid
...
    InternalServerError:
      description: |-
        The server encountered an internal error.
        Further info in server logs.
  parameters:
...
    UserID:
      schema:
        description: Unique identifier of a user
        type: integer
        format: int32
        readOnly: true
      name: userId
      in: path
      required: true
      description: Unique identifier of a user
      example: 1234
  schemas:
...
    UserID:
      description: Unique identifier of a user
      type: integer
      format: int32
      readOnly: true
    BaseUser:
      description: Provides basic information about someone with a WASA Photo account.
      type: object
      required:
        - id
        - username
      properties:
        id:
          $ref: "#/components/schemas/UserID"
        username:
          type: string
          description: User name
          pattern: '^.*?$'
          example: Mario
          minLength: 3
          maxLength: 16
    User:
      allOf:
        - $ref: "#/components/schemas/BaseUser"
        - description: Provides extended information about someone with a WASA Photo account, like total photos count etc.
          type: object
          properties:
            bannedForUser:
              type: boolean
              description: Either if a user is banned or not for the logged-in user
              example: false
            totalPhotos:
              type: integer
              description: Total photos count
              format: int32
              example: 20
            totalFollowers:
              type: integer
              description: Total followers count
              format: int32
              example: 10
            totalFollowings:
              type: integer
              description: Total followings count
              format: int32
              example: 10
...
  links:
...
    GetUserById:
      operationId: getUserProfile
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `GET /users/{userId}` to get complete information about a user
    FollowUser:
      operationId: followUser
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `PUT /users/me/followings/{targetUserId}` to follow the target user
    UnfollowUser:
      operationId: unfollowUser
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `PUT /users/me/followings/{targetUserId}` to unfollow the target user
    BanUser:
      operationId: banUser
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `PUT /users/me/bans/{targetUserId}` to ban the target user
    UnbanUser:
      operationId: unbanUser
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `PUT /users/me/bans/{targetUserId}` to unban the target user
    GetUserPhotos:
      operationId: getUserProfile
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `GET /users/{userId}` to get complete information about a user
    GetUserFollowers:
      operationId: getUserFollowers
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `GET /users/{userId}` to get complete information about a user
    GetUserFollowings:
      operationId: getUserFollowings
      parameters:
        userId: $response.body#id
      description: |-
        The `id` value returned in the response can be used as the `userId` parameter in `GET /users/{userId}` to get complete information about a user
security:
  - bearerAuth: []
paths:
...
  /users:
    get:
      tags: ["Users Lookup"]
      operationId: getUsers
      summary: Get users list
      description: |-
        Get a list of WASA Photo users
      parameters:
        - in: query
          name: username
          schema:
            description: Query param with searched username value
            type: string
            pattern: '^.*?$'
            example: Mario
            minLength: 3
            maxLength: 16
          description: The username to filter users against
      responses:
        "200":
          description: List of user resources
          content:
            application/json:
              schema:
                description: List of user resources
                type: array
                items:
                  $ref: "#/components/schemas/BaseUser"
                minItems: 0
                maxItems: 99999
          links:
            getUserById:
              $ref: "#/components/links/GetUserById"
            followUser:
              $ref: "#/components/links/FollowUser"
            unfollowUser:
              $ref: "#/components/links/UnfollowUser"
            banUser:
              $ref: "#/components/links/BanUser"
            unbanUser:
              $ref: "#/components/links/UnbanUser"
            getUserPhotos:
              $ref: "#/components/links/GetUserPhotos"
            getUserFollowers:
              $ref: "#/components/links/GetUserFollowers"
            getUserFollowings:
              $ref: "#/components/links/GetUserFollowings"
        "500":
          $ref: "#/components/responses/InternalServerError"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/me:
    # get:
    #   tags: ["Users Lookup"]
    #   operationId: getMyUserProfile
    #   summary: Get the current user
    #   description: |-
    #     The resource for the user shows the user id and the username
    #   responses:
    #     "200":
    #       description: A user resource
    #       content:
    #         application/json:
    #           schema:
    #             $ref: "#/components/schemas/User"
    #       links:
    #         getUserPhotos:
    #           $ref: "#/components/links/GetUserPhotos"
    #         getUserFollowers:
    #           $ref: "#/components/links/GetUserFollowers"
    #         getUserFollowings:
    #           $ref: "#/components/links/GetUserFollowings"
    #     "500":
    #       $ref: "#/components/responses/InternalServerError"
    #     '401':
    #       $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: ["Manage Users"]
      operationId: setMyUserName
      summary: Update current username
      description: |-
        Send a new user username that will replace the current one for the user
      requestBody:
        content:
          application/json-patch+json:
            schema:
              description: A patch object containing a replace entry for the username
              type: array
              items:
                type: object
                properties:
                  op:
                    type: string
                    description: Patch operation type
                    enum: ["replace"]
                    example: replace
                    pattern: '^.*?$'
                    minLength: 7
                    maxLength: 7
                  path:
                    type: string
                    description: Field to patch
                    enum: ["/username"]
                    example: /username
                    pattern: '^.*?$'
                    minLength: 9
                    maxLength: 9
                  value:
                    type: string
                    description: New field valued
                    example: Luigi
                    pattern: '^.*?$'
                    minLength: 3
                    maxLength: 16
              minItems: 1
              maxItems: 1
              uniqueItems: true
        required: true
      responses:
        "200":
          description: Username updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "500":
          $ref: "#/components/responses/InternalServerError"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /users/{userId}:
    parameters:
      - $ref: "#/components/parameters/UserID"
    get:
      tags: ["Users Lookup"]
      operationId: getUserProfile
      summary: Get a user
      description: |-
        The resource for the user shows the user id and the username
      responses:
        "200":
          description: A user resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
          links:
            getUserPhotos:
              $ref: "#/components/links/GetUserPhotos"
            getUserFollowers:
              $ref: "#/components/links/GetUserFollowers"
            getUserFollowings:
              $ref: "#/components/links/GetUserFollowings"
            followUser:
              $ref: "#/components/links/FollowUser"
            unfollowUser:
              $ref: "#/components/links/UnfollowUser"
            banUser:
              $ref: "#/components/links/BanUser"
            unbanUser:
              $ref: "#/components/links/UnbanUser"
        "404":
          description: User not found
        "500":
          $ref: "#/components/responses/InternalServerError"
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...